marshal standard inline wrapper
