--- crates/consensus/src/consensus/application/ingress.rs
@@ -1,145 +1,0 @@
-use commonware_actor::Feedback;
-use commonware_consensus::{
-    Automaton, CertifiableAutomaton, Relay,
-    simplex::{Plan, types::Context},
-    types::{Round, View},
-};
-
-use commonware_cryptography::ed25519::PublicKey;
-use commonware_utils::channel::oneshot;
-use futures::{SinkExt as _, channel::mpsc};
-use std::time::Instant;
-
-use crate::consensus::Digest;
-
-#[derive(Clone)]
-pub(crate) struct Mailbox {
-    inner: mpsc::Sender<Message>,
-}
-
-impl Mailbox {
-    pub(super) fn from_sender(inner: mpsc::Sender<Message>) -> Self {
-        Self { inner }
-    }
-}
-
-/// Messages forwarded from consensus to application.
-// TODO: add trace spans into all of these messages.
-pub(super) enum Message {
-    Broadcast(Box<Broadcast>),
-    Propose(Box<Propose>),
-    Verify(Box<Verify>),
-}
-
-pub(super) struct Propose {
-    pub(super) parent: (View, Digest),
-    pub(super) response: oneshot::Sender<Digest>,
-    pub(super) round: Round,
-    pub(super) leader: PublicKey,
-    pub(super) started_at: Instant,
-}
-
-impl From<Propose> for Message {
-    fn from(value: Propose) -> Self {
-        Self::Propose(Box::new(value))
-    }
-}
-
-pub(super) struct Broadcast {
-    pub(super) digest: Digest,
-    pub(super) plan: Plan<PublicKey>,
-}
-
-impl From<Broadcast> for Message {
-    fn from(value: Broadcast) -> Self {
-        Self::Broadcast(Box::new(value))
-    }
-}
-
-pub(super) struct Verify {
-    pub(super) parent: (View, Digest),
-    pub(super) payload: Digest,
-    pub(super) proposer: PublicKey,
-    pub(super) response: oneshot::Sender<bool>,
-    pub(super) round: Round,
-}
-
-impl From<Verify> for Message {
-    fn from(value: Verify) -> Self {
-        Self::Verify(Box::new(value))
-    }
-}
-
-impl Automaton for Mailbox {
-    type Context = Context<Self::Digest, PublicKey>;
-
-    type Digest = Digest;
-
-    async fn propose(&mut self, context: Self::Context) -> oneshot::Receiver<Self::Digest> {
-        // XXX: Cannot propagate the error upstream because of the trait def.
-        // But if the actor no longer responds the application is dead.
-        let (tx, rx) = oneshot::channel();
-        self.inner
-            .send(
-                Propose {
-                    parent: context.parent,
-                    response: tx,
-                    round: context.round,
-                    leader: context.leader,
-                    started_at: Instant::now(),
-                }
-                .into(),
-            )
-            .await
-            .expect("application is present and ready to receive proposals");
-        rx
-    }
-
-    async fn verify(
-        &mut self,
-        context: Self::Context,
-        payload: Self::Digest,
-    ) -> oneshot::Receiver<bool> {
-        // XXX: Cannot propagate the error upstream because of the trait def.
-        // But if the actor no longer responds the application is dead.
-        let (tx, rx) = oneshot::channel();
-        self.inner
-            .send(
-                Verify {
-                    parent: context.parent,
-                    payload,
-                    proposer: context.leader,
-                    round: context.round,
-                    response: tx,
-                }
-                .into(),
-            )
-            .await
-            .expect("application is present and ready to receive verify requests");
-        rx
-    }
-}
-
-// TODO: figure out if this can be useful for tempo. The original PR implementing
-// this trait:
-// https://github.com/commonwarexyz/monorepo/pull/2565
-// Associated issue:
-// https://github.com/commonwarexyz/monorepo/issues/1767
-impl CertifiableAutomaton for Mailbox {
-    // NOTE: uses the default impl for CertifiableAutomaton which always
-    // returns true.
-}
-
-impl Relay for Mailbox {
-    type Digest = Digest;
-    type PublicKey = PublicKey;
-    type Plan = commonware_consensus::simplex::Plan<PublicKey>;
-
-    fn broadcast(&mut self, digest: Self::Digest, plan: Self::Plan) -> Feedback {
-        match self.inner.try_send(Broadcast { digest, plan }.into()) {
-            Ok(()) => Feedback::Ok,
-            Err(error) if error.is_full() => Feedback::Backoff,
-            Err(_) => Feedback::Closed,
-        }
-    }
-}
