+++ crates/consensus/src/epoch/manager/mod.rs
@@ -4,10 +4,14 @@ pub(super) mod ingress;
 use std::{sync::Arc, time::Duration};
 
 pub(crate) use actor::Actor;
+use commonware_consensus::{
+    CertifiableAutomaton, Relay,
+    simplex::{Plan, types::Context},
+    types::{FixedEpocher, ViewDelta},
+};
 use commonware_cryptography::ed25519::PublicKey;
 pub(crate) use ingress::Mailbox;
 
-use commonware_consensus::types::{FixedEpocher, ViewDelta};
 use commonware_p2p::Blocker;
 use commonware_runtime::{
     BufferPooler, Clock, Metrics, Network, Spawner, Storage, buffer::paged::CacheRef,
@@ -15,10 +19,10 @@ use commonware_runtime::{
 use rand_08::{CryptoRng, Rng};
 use tempo_node::TempoFullNode;
 
-use crate::{epoch::scheme_provider::SchemeProvider, feed, subblocks};
+use crate::{consensus::Digest, epoch::scheme_provider::SchemeProvider, feed, subblocks};
 
-pub(crate) struct Config<TBlocker> {
-    pub(crate) application: crate::consensus::application::Mailbox,
+pub(crate) struct Config<TBlocker, TAutomaton> {
+    pub(crate) automaton: TAutomaton,
     pub(crate) execution_node: Arc<TempoFullNode>,
     pub(crate) blocker: TBlocker,
     pub(crate) page_cache: CacheRef,
@@ -37,12 +41,14 @@ pub(crate) struct Config<TBlocker> {
     pub(crate) views_until_leader_skip: ViewDelta,
 }
 
-pub(crate) fn init<TContext, TBlocker>(
+pub(crate) fn init<TContext, TBlocker, TAutomaton>(
     context: TContext,
-    config: Config<TBlocker>,
-) -> (Actor<TContext, TBlocker>, Mailbox)
+    config: Config<TBlocker, TAutomaton>,
+) -> (Actor<TContext, TBlocker, TAutomaton>, Mailbox)
 where
     TBlocker: Blocker<PublicKey = PublicKey>,
+    TAutomaton: CertifiableAutomaton<Context = Context<Digest, PublicKey>, Digest = Digest>
+        + Relay<Digest = Digest, PublicKey = PublicKey, Plan = Plan<PublicKey>>,
     TContext: BufferPooler
         + Spawner
         + Metrics
